\name{assembleFeatureMatrix}
\title{assembleFeatureMatrix}
\description{Assemble outcome and prediction features for a TF}
\usage{
assembleFeatureMatrix(tf)
}
\arguments{
  \item{tf}{the symbol of the TF. muse be in hg38_train.}
}
\value{ data frame with footprints in rows and features in columns.}
\author{ Seth Ament }




